server/api/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
h11==0.14.0
supabase==0.7.1
httpcore>=0.15.0,<0.17.0
httplib2==0.22.0
httpx==0.23.3
idna>=2.8,<3.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
lightgbm==4.6.0
MarkupSafe==3.0.2
numpy==2.2.3
pandas==2.2.3
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic>=1.9.0,<2.0.0
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.25.1
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.5.0
tqdm==4.67.1
python-dotenv==0.19.0
typing_extensions==4.12.2
flask-cors==4.0.0
tzdata==2025.1
uritemplate==4.1.1
urllib3>=1.21.1,<1.27
Werkzeug==3.1.3
xgboost==2.1.4
